close #73: make try/catch work for doFuture.
Note: This commit also changes the internals of Future execution:
1. No longer will doFuture() eagerly execute code until the first yield.
2. Instead of creating a chain of Futures, doFuture will be executed sequentially
   in the fork() state machine.
3. We moved back to computes and now use promises internally to execute Run code.
4. Future chains should finally be stack safe.

Note that due to the way fork() now works, it is not possible to catch a Raise returned
in a generator. This is because the generators ends before we can process it. Store final
values in a variable first and then return them.